TASK 7

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -4,7 +4,7 @@ DROP PROCEDURE IF EXISTS ComputeAverageScoreForUser;
 DELIMITER $$
 CREATE PROCEDURE ComputeAverageScoreForUser(IN user_id INT)
 BEGIN
-	DECLARE avg_score FLOAT;
+	DECLARE avg_score DECIMAL(10, 2);
 	SELECT AVG(score) INTO avg_score
 	FROM corrections
 	WHERE user_id = user_id;
